requirements.txt
Flask==3.0.2
waitress==3.0.1
requests==2.31.0
PyMuPDF==1.26.0
Pillow==10.4.0
pandas==2.2.3
numpy==1.26.4
xlwt==1.3.0

google-cloud-documentai==3.5.0
google-cloud-secret-manager==2.23.3
google-cloud-storage==2.19.0
google-cloud-aiplatform>=1.75.0

